codec: mixed-wire-order gate fixtures + honor value-class wireOrder + drop redundant casts
Three changes in one commit because the fixture surfaced the value-class
bug and exercising the new emit path required regenerated snapshots:

1. **Mixed-wire-order gate fixtures.** Three new fixtures pin the
   coalescer's flushing behaviour when adjacent batchable fields
   resolve to different wire orders:
   - `MixedOrderFlush` — parent Big, middle field overrides to Little
     via `@WireOrder`. Three separate emits expected.
   - `MixedOrderValueClass` — parent Big, value-class field with its
     own `wireOrder=Little`. Same three-emit shape, but flushing on
     `field.valueClassWireOrder`.
   - `MixedOrderPartialBatch` — parent Big with two Big fields then a
     Little-override field. First two batch into one readInt; third
     emits alone. Catches both an under-flushing gate (wrong middle
     bytes) and an over-flushing gate (no batch on the first pair).
   12 round-trip tests with exact wire-byte assertions across both
   buffer orders.

2. **Honor value-class wireOrder in single-field emit.** The
   pre-existing `appendDecode/EncodeValueClassScalar` used
   `naturalScalarReadExpr` / `naturalScalarWriteStatement`, which
   trust buffer.byteOrder and ignore the value class's own declared
   wireOrder. The new `MixedOrderValueClass` fixture surfaced this:
   the gate flushed on `valueClassWireOrder = Little`, but the
   resulting single-field emit ignored Little and used buffer order
   — producing wrong bytes under a mismatching buffer. Both
   functions now route multi-byte explicit-order value classes
   through `appendValueClassNaturalReadWithSwap` /
   `appendValueClassNaturalWriteWithSwap`, matching the contract
   the docs already promise: explicit wireOrder beats buffer order.
   `LeHeader` (the existing pre-#154 value class) gains the swap;
   `LePacket` tests still pass because they pair LE buffer with LE
   value class (the no-swap arm).

3. **Drop redundant `.toInt()` / `.toLong()` casts in batched
   encode.** `convertBatchAccumulatorForWrite` wrapped the already-
   correct-type combined expression in a redundant conversion for
   size 4/8. Compiler emitted "Redundant call of conversion method"
   warnings on every batched-encode site. Size-2 still narrows Int
   to Short for writeShort.

All cross-platform tests green: JVM, JS, Linux native, WASM. Snapshot
regen lockstep with the value-class wireOrder fix.

+++ buffer-codec-test/codec-snapshots/com/ditchoom/buffer/codec/test/protocols/batch/BigHeaderCodec.kt
@@ -28,7 +28,7 @@ public object BigHeaderCodec : Codec<BigHeader> {
     `value`: BigHeader,
     context: EncodeContext,
   ) {
-    val __batch2 = (((value.type.toLong() and 0xFFL) shl 56) or ((value.version.toLong() and 0xFFL) shl 48) or ((value.flags.toLong() and 0xFFFFL) shl 32) or (value.length.toLong() and 0xFFFFFFFFL)).toLong()
+    val __batch2 = ((value.type.toLong() and 0xFFL) shl 56) or ((value.version.toLong() and 0xFFL) shl 48) or ((value.flags.toLong() and 0xFFFFL) shl 32) or (value.length.toLong() and 0xFFFFFFFFL)
     buffer.writeLong(if (buffer.byteOrder == ByteOrder.BIG_ENDIAN) __batch2 else swapBytes(__batch2))
   }
